driver: rework RX FIFO-full handling logic
Previously the driver tried to prevent any RX FIFO full IRQ by eventually
discarding the older DMA buffers (counting the so-called "pktloss" events).

In case a RX FIFO full event happened anyway then a weird force-close
behaviour made the read() fail and brought down the whole thing.

All this mechanism seems quite insane: there is no point in avoidind a FIFO-
full event by deliberately throwing away data; and it was even utterly
buggy, causing kernel OOPS sometimes..

This patch reworks completely the RX FIFO-full handling logic:

First of all we do not struggle in avoiding a FIFO-full by quickly throwing
away old buffers: this would avoid the HPU-core raising erros, but it wouldn't
prevent any data loss neither bring any other advantage. So in case the reader
is not fast enough, or the DMA is unable to unload the HPU fifo, we just let
the HPU to raise the FIFO-full error.

In those condition we still let the ongoing/next(!) read() to return error, so
that the user is aware of this condition. Internally, we stop feeding the HPU
FIFO and we start to drain data from the DMA buffers (side effet: the HPU
FIFO gets also unloaded) so that we can restart in a somehow clean situation
(some leftovers might actually survive).

As soon as a new read() is issued (after the one that returned error), the RX
mechanism is restored.