Fix fallthrough warnings
The Linux kernel now (for a couple of years already) requires
fall-through from one case statement to the next to be annotated with
the `fallthrough` macro, so put this in instead of /* fall through */
comments.

To continue to allow compilation against older kernels that don't have the
fallthrough macro, define it away if it's not defined.