iit-hpucore-dma.c: Rename debugfs register file names to match doc
Similarly to 476b2281 which renamed the register #define's to match the 
documentation, this renames the files created in debugfs in the same 
way. This is done as a separate commit to make it easily revertable 
because this change could break programs or scripts that read from these 
files. (None appear to be present in this repository.)