HPU_CTRL_ bit definitions: rename to match documentaion, add missing
The end goal is to export the register and register bit definitions from 
the C .h file to Python to make them usable by users of  Pynq and any 
other platform on which we run Python in the future. To make things 
easier for such users, it will help to name the registers and bit-fields 
as much as possible to be similar to the names given in the hardware 
documentation (in the doc directory). The registers were renamed in 
476b2281, whereas this commit deals with the fields in the HPU_CTRL_REG.